Move geometry extraction to its own file
